requirements.txt
Django>=3.2
djangorestframework>=3.12
django-ckeditor>=6.0.0
django-redis>=5.2.0
googletrans==4.0.0-rc1
pytest-django>=4.0.0
gunicorn
redis
